Renamed Post/PostComponent.tsx to Post/index.tsx. Fixed AdminButtons import
